+++ test/bin.sh
@@ -1,19 +1,36 @@
 #!/bin/bash
-VER=$(curl -s https://api.github.com/repos/telatin/seqfu2/releases/latest  | perl -nE 'my ($tag, $val) = split /:/, $_; if ($tag=~/tag_name/) { my @tag = split /"/, $val; for my $i (@tag) { $i =~s/[^0-9.]//g; say $i if (length($i) > 2); } }')
 
+SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+
+echo "Building fresh binary"
+echo "============================="
+
+VER=$(curl -s https://api.github.com/repos/telatin/seqfu2/releases/latest  | perl -nE '
+my ($tag, $val) = split /:/, $_;
+if ($tag=~/tag_name/) { 
+  my @tag = split /"/, $val; 
+  for my $i (@tag) { 
+    $i =~s/[^0-9.]//g; 
+    say $i if (length($i) > 2); 
+  }
+}')
+
+ 
 if [[ $(uname) == "Darwin" ]]; then
-  PLAT='macos-intel';
+  PLAT='Darwin';
 else
-  PLAT='linux64'
+  PLAT='Linux-x86_64'
 fi
 
-if [[ -d "$HOME/Downloads" ]];
-then
-  DEST="$HOME/Downloads";
-else
-  DEST="./release"
-fi
-
+DEST="$SCRIPT_DIR/../releases/zips/";
 mkdir -p $DEST
-nimble build && zip -r $DEST/seqfu-${VER}-${PLAT}.zip bin/*
 
+echo "Last version online: $VER"
+echo "Local version: $LOCALVER"
+nimble build
+LOCAL_RELEASE=$(./bin/seqfu version)
+zip -r $DEST/SeqFu-v${LOCAL_RELEASE}-${PLAT}.zip bin/*
+
+echo "Last version online: $VER"
+echo "Local version: $LOCAL_RELEASE"
+echo "$DEST/SeqFu-v${LOCAL_RELEASE}-${PLAT}.zip"