+++ releases/changes.md
@@ -1,2 +1,3 @@
 * Updated tests
 * Fixed bug in `fu-orf` to allow for single sequences
+* `fu-orf` also prints frame in the sequence comment